[refactor] 컨트롤러 추가 분리
